requirements.txt
alembic==1.12.1
attrs==23.1.0
blinker==1.7.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
url-normalize==1.4.3
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
alembic==1.12.1
attrs==23.1.0
blinker==1.7.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
url-normalize==1.4.3
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0